added TCP_SHARED class will have all the TCP shared functions in it, working on first implementation of player turn for Online player
