+++ CMakeLists.txt
@@ -27,6 +27,8 @@ add_executable(Zeeslag_Code_Fixing main.cpp
         TCP_Connection/Tcp_Client_Socket.cpp
         TCP_Connection/Tcp_Server_Socket.cpp
         TCP_Connection/Tcp_Client_Socket.h
-        TCP_Connection/Tcp_Server_Socket.h)
+        TCP_Connection/Tcp_Server_Socket.h
+        TCP_Connection/TCP_SHARED.cpp
+        TCP_Connection/TCP_SHARED.h)
 
-target_link_libraries(Data_serialization_TCP PRIVATE nlohmann_json::nlohmann_json)
+target_link_libraries(Zeeslag_Code_Fixing PRIVATE nlohmann_json::nlohmann_json)
