+++ Game_Manager_Related/Game_Manager_Loader.h
@@ -21,7 +21,6 @@ private:
 
     void forward_SP_playgame();
     void forward_MP_playgame();
-
 };
