+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Dec 06 13:56 CET
+Start testing: Dec 06 14:31 CET
 ----------------------------------------------------------
-End testing: Dec 06 13:56 CET
+End testing: Dec 06 14:31 CET
